+++ src/refactoring/components/product/index.ts
@@ -1,2 +1,3 @@
+export { EditProductForm } from './EditProductForm';
 export { NewProductForm } from './NewProductForm';
 export { ProductAccordion } from './ProductAccordion';
